Revert "decouple database config"
